+++ src/account/account-data.ts
@@ -1,23 +1,16 @@
 import { utils, Wallet } from 'ethers'
-import {
-  assertAccount,
-  assertMnemonic,
-  assertPassword,
-  assertUsername,
-  CHUNK_ALREADY_EXISTS_ERROR,
-  HD_PATH,
-  removeZeroFromHex,
-} from './utils'
+import { assertAccount, assertMnemonic, assertPassword, assertUsername, HD_PATH, removeZeroFromHex } from './utils'
 import { getEncryptedMnemonic } from './mnemonic'
 import { decryptText } from '../utils/encryption'
 import { downloadPortableAccount, uploadPortableAccount, UserAccountWithMnemonic } from './account'
 import { Connection } from '../connection/connection'
-import { AddressOptions, isAddressOptions, isMnemonicOptions, MnemonicOptions } from './types'
+import { AddressOptions, isAddressOptions, isMnemonicOptions, MnemonicOptions, RegistrationRequest } from './types'
 import { ENS, PublicKey } from '@fairdatasociety/fdp-contracts-js'
 import { Reference, Utils } from '@ethersphere/bee-js'
 import CryptoJS from 'crypto-js'
 import { bytesToHex } from '../utils/hex'
 import { mnemonicToSeed, prepareEthAddress, privateKeyToBytes } from '../utils/wallet'
+import { isChunkAlreadyExistsError, isInsufficientFundsError } from '../utils/error'
 
 export class AccountData {
   /**
@@ -129,7 +122,7 @@ export class AccountData {
     const exported = await this.exportWallet(username, password, options)
     this.setAccountFromMnemonic(exported.mnemonic)
 
-    return this.register(username, password)
+    return this.register(this.createRegistrationRequest(username, password))
   }
 
   /**
@@ -161,12 +154,32 @@ export class AccountData {
   }
 
   /**
-   * Creates new FDP account and gives back user account with swarm reference
+   * Creates a request object that is used to invoke the 'register' method. This object
+   * encapsulates the complete state of registration process. In case when registration
+   * fails in any of the steps, the 'register' method can be safely invoked again with
+   * the existing RegistrationRequest object. The process will continue from the failed
+   * step.
    *
    * @param username FDP username
    * @param password FDP password
+   *
    */
-  async register(username: string, password: string): Promise<Reference> {
+  createRegistrationRequest(username: string, password: string): RegistrationRequest {
+    return {
+      username,
+      password,
+      ensCompleted: false,
+    }
+  }
+
+  /**
+   * Creates new FDP account and gives back user account with swarm reference
+   *
+   * @param request a RegistrationRequest object that contains the state of registration process
+   */
+  async register(request: RegistrationRequest): Promise<Reference> {
+    const { username, password, ensCompleted } = request
+
     assertUsername(username)
     assertPassword(password)
     assertAccount(this)
@@ -175,7 +188,16 @@ export class AccountData {
 
     try {
       const seed = CryptoJS.enc.Hex.parse(removeZeroFromHex(bytesToHex(this.seed!)))
-      await this.ens.registerUsername(this.ens.createRegisterUsernameRequest(username, wallet.address, this.publicKey!))
+
+      if (!ensCompleted) {
+        if (!request.ensRequest) {
+          request.ensRequest = this.ens.createRegisterUsernameRequest(username, wallet.address, this.publicKey!)
+        }
+
+        await this.ens.registerUsername(request.ensRequest)
+      }
+
+      request.ensCompleted = true
 
       return await uploadPortableAccount(
         this.connection,
@@ -185,10 +207,10 @@ export class AccountData {
         seed,
       )
     } catch (e) {
-      const error = e as Error
-
-      if (error.message?.startsWith(CHUNK_ALREADY_EXISTS_ERROR)) {
+      if (isChunkAlreadyExistsError(e)) {
         throw new Error('User account already uploaded')
+      } else if (isInsufficientFundsError(e)) {
+        throw new Error('Not enough funds')
       } else {
         throw e
       }
@@ -235,9 +257,7 @@ export class AccountData {
         seed,
       )
     } catch (e) {
-      const error = e as Error
-
-      if (!error.message?.startsWith(CHUNK_ALREADY_EXISTS_ERROR)) {
+      if (!isChunkAlreadyExistsError(e)) {
         throw e
       }
     }