+++ src/file/utils.ts
@@ -1,5 +1,5 @@
 import { Connection } from '../connection/connection'
-import { Bee, Reference, RequestOptions, UploadResult } from '@ethersphere/bee-js'
+import { Bee, Reference, BeeRequestOptions, UploadResult } from '@ethersphere/bee-js'
 import { PathInfo } from '../pod/utils'
 import { Blocks, FileShareInfo, RawBlock, RawBlocks } from './types'
 import { rawBlocksToBlocks } from './adapter'
@@ -9,7 +9,6 @@ import { FileMetadata, RawFileMetadata } from '../pod/types'
 import { EncryptedReference } from '../utils/hex'
 import { isRawFileMetadata, splitPath } from '../directory/utils'
 import { getUnixTimestamp } from '../utils/time'
-import { bytesToString } from '../utils/bytes'
 import { jsonParse } from '../utils/json'
 
 /**
@@ -110,10 +109,10 @@ export function getBaseName(path: string): string | undefined {
 export async function downloadBlocksManifest(
   bee: Bee,
   reference: Reference,
-  downloadOptions?: RequestOptions,
+  downloadOptions?: BeeRequestOptions,
 ): Promise<Blocks> {
-  const data = await bee.downloadData(reference, downloadOptions)
-  const rawBlocks = jsonParse(bytesToString(data), 'blocks manifest')
+  const data = (await bee.downloadData(reference, downloadOptions)).text()
+  const rawBlocks = jsonParse(data, 'blocks manifest')
   assertRawBlocks(rawBlocks)
 
   return rawBlocksToBlocks(rawBlocks)